fix: schema delete quote

+++ api/resources/quote.py
@@ -59,10 +59,15 @@ class QuoteResource(Resource):
         db.session.commit()
         return quote.to_dict(), 200
 
-    def delete(self, quote_id):
+    def delete(self, author_id, quote_id):
+        author = AuthorModel.query.get(author_id)
+        if author is None:
+            return {"Error": f"Author id={author_id} not found"}, 404
         quote = QuoteModel.query.get(quote_id)
         if quote is None:
             return {"Error": f"Quote id={quote_id} not found"}, 404
+        if quote.author.id != author.id:
+            return {"Error": "Цитата не принадлежит автору"}, 400
         db.session.delete(quote)
         db.session.commit()
         return "", 204
